chore(models): keep asset implementation within limits
